Changed to only display data packets, and to ignore ack packets that dont contain data.
Just noticed that there should be more checking as to what protocol is
being used before things are done. It works how it is now, but I think
part of that is down to 'random' numbers in parts of the packet that arent
close to values that could be legit. This is a commit so I can rip all that
up without breaking anything that is already working.