Changed the module API (again).
Modules now get "init_module()" called with the parameters passed from the
config file.  The position modules init functions also get passed if they
are left or right.  if the init_module function doens't exist then it is
not called.

the position module api now passes which side it's on (left/right), and
if the packet is inbound/outbound which it can use to tell which way the
packet is travelling (there is a CHK_SOURCE macro to help here).

And end_module function is also called if it exists to cleanup after
init_module.

This is designed so that users can pass arbitary parameters to arbitary
modules now